Crate wayland_csd_frame

The interface for wayland client side decorations (CSD).

The crate is intended to be used by libraries providing client side decorations for the xdg-shell protocol.

Examples could be found in client toolkit and sctk-adwaita.

Structs

CursorIconParseError

An error which could be returned when parsing CursorIcon.

WindowManagerCapabilities

The capabilities of the window manager.

WindowState

The configured state of the window.

Enums

CursorIcon

Describes the appearance of the (usually mouse) cursor icon.

FrameAction

The Frame action user should perform in responce to mouse click events.

FrameClick

The user clicked or touched the decoractions frame.

ResizeEdge

Which edge or corner is being dragged.

Traits

DecorationsFrame

The interface for the client side decorations.